Move bindings to its own class
