+++ Engine/Bindings.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: ef4c73daf9b556744894ef45ad6cae19
+timeCreated: 1515613506
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
