+++ Examples/RumorScriptExample.cs
@@ -66,8 +66,8 @@ label end:
 ");
 			rumor.Scope.DefaultSpeaker = "Narrator";
 
-			rumor.Bind("get_apples", () => { return Random.Range(2, 6); });
-			rumor.Bind("get_pears", () => { return Random.Range(2, 6); });
+			rumor.Bindings.Bind("get_apples", () => { return Random.Range(2, 6); });
+			rumor.Bindings.Bind("get_pears", () => { return Random.Range(2, 6); });
 
 			StartCoroutine(rumor.Run());
 		}
