+++ README.md
@@ -89,8 +89,8 @@ public class RumorScriptExample : MonoBehaviour
         rumor = new Rumor(File.ReadAllText("Example.txt"));
         rumor.Scope.DefaultSpeaker = "Narrator";
 
-        rumor.Bind("get_apples", () => { return Random.Range(2, 6); });
-        rumor.Bind("get_pears", () => { return Random.Range(2, 6); });
+        rumor.Bindings.Bind("get_apples", () => { return Random.Range(2, 6); });
+        rumor.Bindings.Bind("get_pears", () => { return Random.Range(2, 6); });
 
         StartCoroutine(rumor.Run());
     }
